Objectives:  DONE The application must be developed in Kotlin and the interface must be Jetpack Compose.
Develop an Android application that works on both mobile and Android XR.
The application displays a list of videos, retrieved from an API, and then plays a video when the user clicks on one.

Objectives:  DONE
On Android XR, the video must be displayed in a new window, separate from the video list. A 3D model of a soccer field must appear in front of the video window when the video starts playing.

TODO :

Cosmetics